docker_assignment/backend/requirements.txt
tensorflow
pillow
uvicorn
fastapi
requests
pydantic
matplotlib
starlette
